+++ Runtime/Events/EventsManager.cs
@@ -7,6 +7,7 @@ namespace AffiseAttributionLib.Events
     internal class EventsManager
     {
         private const long TIME_SEND_REPEAT = 15 * 1000;
+        private const long TIME_SEND_SECOND = 3 * 1000;
         private const long SCHEDULER_SEND_REPEAT = 15 * 1000;
 
         private readonly ISendDataToServerUseCase _sendDataToServerUseCase;
@@ -36,11 +37,22 @@ namespace AffiseAttributionLib.Events
             //Send events on start
             SendEvents(false);
             
+            //Send events on delay
+            SendEventsDelay(delay: TIME_SEND_SECOND);
+                
             //Start timer fo repeat send events
             StartTimer();
             
             //Start Scheduler
             StartQueue();
+        }
+
+        private void SendEventsDelay(long delay)
+        {
+            _executorServiceProvider.ExecuteWithDelay(delay, () =>
+            {
+                SendEvents(withDelay: false);
+            });
         }
 
         private void StartTimer()