+++ Runtime/Module/Attribution/IAffiseAttributionModuleApi.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using AffiseAttributionLib.Module.Advertising;
 using AffiseAttributionLib.Module.AppsFlyer;
+using AffiseAttributionLib.Module.Google;
 using AffiseAttributionLib.Module.Link;
 using AffiseAttributionLib.Module.Subscription;
 using AffiseAttributionLib.Module.TikTok;
@@ -16,6 +17,7 @@ namespace AffiseAttributionLib.Module.Attribution
         public IAffiseModuleAdvertisingApi Advertising { get; }
         public IAffiseModuleLinkApi Link { get; }
         public IAffiseModuleAppsFlyerApi AppsFlyer { get; }
+        public IAffiseModuleGoogleApi Google { get; }
         public IAffiseModuleSubscriptionApi Subscription { get; }
         public IAffiseModuleTikTokApi TikTok { get; }