+++ Runtime/Module/Google/AffiseGoogle.cs
@@ -1,0 +1,12 @@
+#nullable enable
+using AffiseAttributionLib.Modules;
+
+namespace AffiseAttributionLib.Module.Google
+{
+    internal class AffiseGoogle : AffiseModuleApiWrapper<IAffiseGoogleApi>, IAffiseModuleGoogleApi
+    {
+        protected override AffiseModules Module => AffiseModules.Google;
+
+        public bool HasModule() => IsModuleInit;
+    }
+}