+++ Runtime/Module/Google/GoogleModule.cs
@@ -1,0 +1,12 @@
+#nullable enable
+using AffiseAttributionLib.Modules;
+
+namespace AffiseAttributionLib.Module.Google
+{
+    public class GoogleModule : AffiseModule, IAffiseGoogleApi
+    {
+        public override void Start()
+        {
+        }
+    }
+}