+++ Runtime/Module/Google/IAffiseGoogleApi.cs
@@ -1,0 +1,9 @@
+#nullable enable
+using AffiseAttributionLib.Modules;
+
+namespace AffiseAttributionLib.Module.Google
+{
+    public interface IAffiseGoogleApi : IAffiseModuleApi
+    {
+    }
+}