+++ Runtime/Module/Google/IAffiseModuleGoogleApi.cs
@@ -1,0 +1,8 @@
+#nullable enable
+
+namespace AffiseAttributionLib.Module.Google
+{
+    public interface IAffiseModuleGoogleApi : IAffiseGoogleApi, IAffiseHasModule
+    {
+    }
+}