+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [1.7.8] - 2026-06-16
+
+### Added
+
+- iOS New module `Google` 
+- iOS Provider `odm_info`
+
+### Changed
+
+- Initial data send
+- Update native iOS to [`1.7.8`](https://github.com/affise/affise-mmp-sdk-ios/blob/1.7.8/CHANGELOG.md)
+- Update native Android to [`1.7.8`](https://github.com/affise/affise-mmp-sdk-android/blob/v1.7.8/CHANGELOG.md)
+
 ## [1.7.7] - 2026-05-19
 
 ### Added
@@ -564,6 +577,8 @@
 - Update native iOS to `1.6.20`
 - Update native Android to `1.6.22`
 
+[1.7.8]: https://github.com/affise/affise-mmp-sdk-unity/compare/1.7.7...1.7.8
+[1.7.7]: https://github.com/affise/affise-mmp-sdk-unity/compare/1.7.6...1.7.7
 [1.7.6]: https://github.com/affise/affise-mmp-sdk-unity/compare/1.7.5...1.7.6
 [1.7.5]: https://github.com/affise/affise-mmp-sdk-unity/compare/1.7.4...1.7.5
 [1.7.4]: https://github.com/affise/affise-mmp-sdk-unity/compare/1.7.3...1.7.4